+++ R/tailr-transforms.R
@@ -1,0 +1,11 @@
+
+#' @export
+transform_call.tailr_pmatch_cases <- function(fun, expr) {
+    stopifnot(rlang::call_name(expr) == "cases")
+
+    args <- rlang::call_args(expr)
+    value <- args[[1]]
+    patterns <- args[-1]
+    eval(rlang::expr(cases_expr(!!value, !!!patterns)))
+}
+