+++ NEWS.md
@@ -1,5 +1,6 @@
 # pmatch 0.1.2
 
+ * Transformation rules to make pattern matching work with tail-recursion from tailr package.
  * Function cases_expr generates an expression (a call object) based on pattern rules.
  * Exporting test_pattern and test_pattern_ to test one pattern against a value.